ptp: Apply missing changes to support > 4GB transfers.
Not sure why, but these changes were dropped when the original
patchset was applied. The rest of the changes to ptp made it in.

+++ src/libopenusb1-glue.c
@@ -1202,7 +1202,7 @@ ptp_usb_senddata(PTPParams* params, PTPContainer* ptp,
     int wlen, datawlen;
     unsigned long written;
     PTPUSBBulkContainer usbdata;
-    uint32_t bytes_left_to_transfer;
+    uint64_t bytes_left_to_transfer;
     PTPDataHandler memhandler;
 
     LIBMTP_USB_DEBUG("SEND DATA PHASE\n");
